+++ Common/Crosshairs/ItemCrosshairController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using Terraria;
+using Terraria.ID;
 using Terraria.ModLoader;
 using TerrariaOverhaul.Core.ItemComponents;
 using TerrariaOverhaul.Core.Time;
@@ -10,9 +11,35 @@ namespace TerrariaOverhaul.Common.Crosshairs;
 [Autoload(Side = ModSide.Client)]
 public sealed class ItemCrosshairController : ItemComponent
 {
+	public override void SetDefaults(Item item)
+	{
+		static bool CheckItem(Item item)
+		{
+			if (item.shoot <= ProjectileID.None) {
+				return false;
+			}
+
+			if (item.buffType > 0) {
+				return false;
+			}
+
+			if (item.pick > 0 || item.axe > 0 || item.hammer > 0) {
+				return false;
+			}
+
+			return false;
+		}
+
+		Enabled |= CheckItem(ContentSamples.ItemsByType.TryGetValue(item.type, out var baseItem) ? baseItem : item);
+	}
+
 	public override bool? UseItem(Item item, Player player)
 	{
 		if (!Enabled || !player.IsLocal()) {
+			return null;
+		}
+
+		if (item.channel) {
 			return null;
 		}
 
@@ -28,6 +55,10 @@ public sealed class ItemCrosshairController : ItemComponent
 	public override void UseAnimation(Item item, Player player)
 	{
 		if (!Enabled || !player.IsLocal()) {
+			return;
+		}
+
+		if (item.channel) {
 			return;
 		}